progress[major] - added `text_handling.py` script
