+++ docs/BUILDING.md
@@ -58,6 +58,7 @@ source/dialog_handling.py \
 source/enemy_handling.py \
 source/data_handling.py \
 source/npc_handling.py \
+source/text_handling.py \
 --add-data yamale/VERSION:yamale \
 --collect-submodules fsspec \
 --collect-submodules appdirs \
